github pages で動作するように変更
